AI-Project-Showcase/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 模型相关
onnx>=1.14.0
onnxruntime>=1.16.0
tensorrt>=8.5.0  # 可选，用于 GPU 优化

# 机器学习库
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0

# 实验跟踪
wandb>=0.15.0
mlflow>=2.8.0

# 图像处理
albumentations>=1.3.0
imageio>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API 和部署
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 开发工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0

# 配置文件
pyyaml>=6.0.1
omegaconf>=2.3.0

# 日志和进度
tqdm>=4.66.0
loguru>=0.7.0

# 其他工具
timm>=0.9.0
einops>=0.7.0
